Split autoloader from autogen
